Move OnAudioFocusChangeListener to jinterface
